requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
install==1.3.4
mysqlclient==2.0.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
